requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==1.4.41
pandas==2.1.4
pandera==0.18.0
multimethod==1.9.1
python-dotenv==1.0.0
faker==22.0.0
pytest==8.1.1
pytest-cov==4.1.0
greenlet==3.0.3